fix: rename crudDatabase/test to crudDatabase/testResult since it's conflicted with unit test
